+++ src/ziggurat/sentry.clj
@@ -1,4 +1,5 @@
 (ns ziggurat.sentry
+  "Sentry reporting functions used by both Ziggurat and the actor."
   (:require [clojure.tools.logging :as log]
             [ziggurat.config :refer [ziggurat-config]]
             [mount.core :refer [defstate]]
@@ -21,5 +22,8 @@
           (sentry/shutdown-reporter sentry-reporter)))
 
 (defmacro report-error
+  "Logs an exception and reports it to Sentry.
+  Create an exception using ex-info if you don't have an exception but
+  still want to report to Sentry."
   [^Throwable error & msgs]
   `(sentry/report-error sentry-reporter ~error ~@msgs))